Removed 'community' sentence in WordPress Plugins section
I've suggested removing the last sentence as the other ones don't have any mention about the community, so they are all consistent. Maybe the description could be improved as well, I especially found this sentence "Most of the Frontity plugins will extend the REST API so it makes easier to connect the frontend" a bit confusing.